requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-maintenance-mode==0.16.3
gunicorn==20.1.0
python-dotenv==0.20.0
requests==2.27.1